Use Float16 to save memory.

+++ pysegcnn/core/dataset.py
@@ -634,7 +634,7 @@ class ImageDataset(Dataset):
 
         # initialize dictionary of class spectral distribution
         # exclude NoData class
-        cls_ds = {k: np.empty(shape=(0, len(self.use_bands)), dtype=np.float32)
+        cls_ds = {k: np.empty(shape=(0, len(self.use_bands)), dtype=np.float16)
                   for k, v in self.labels.items() if v['label'] != 'No_data'}
 
         # iterate over the samples of the dataset
